daily-brief/requirements.txt
PyYAML>=6.0
feedparser>=6.0
python-dotenv>=1.0.0
openai>=1.0.0
